player death working
