+++ Assets/Player/GasManager.cs
@@ -9,9 +9,14 @@ public class GasManager : MonoBehaviour
     [SerializeField] float currentGas;
     [SerializeField] float gasRemoveRate;
     [SerializeField] Slider gasSlider;
-    [SerializeField] GameObject failText;
 
-    public float CurrentGas { get => currentGas; set => currentGas = value; }
+    private GameManager _gameManager;
+
+    public float CurrentGas
+    {
+        get => currentGas;
+        set => currentGas = value;
+    }
 
     void Update()
     {
@@ -35,7 +40,7 @@ public class GasManager : MonoBehaviour
                 magnet.PulledMagnetsInField.Remove(GetComponent<Magnet>());
             }
 
-            failText.SetActive(true);
+            FindObjectOfType<GameManager>().FailGame();
         }
     }
 }
